+++ src/apiError.js
@@ -1,18 +1,19 @@
 /**
  *
  * @export
- * @class ApiError
+ * @class AuthressHttpError
  * @extends {Error}
  */
-class ApiError extends Error {
-  constructor(message, status, body, headers) {
-    super(message);
-    this.name = 'ApiError';
-    this.code = 'ApiError';
+class AuthressHttpError extends Error {
+  constructor(url, status, data, headers) {
+    super(data?.title || url);
+    this.name = 'AuthressHttpError';
+    this.code = 'AuthressHttpError';
+    this.url = url;
     this.status = status;
-    this.body = body;
+    this.data = data;
     this.headers = headers;
   }
 }
 
-module.exports = ApiError;
+module.exports = AuthressHttpError;